Training pipeline

Data parallel training happens over several GPUs which themselves may be spread across sevral nodes. While there are several architectures proposed for high throughput and fast training of neural networks with a large amount of data, it is out of scope for this document. Here we will focus on the tooling necessary to achieve training over several GPUs and the API.

Syncing Gradients

ResNetImageNet.syncgrads — Function

syncgrads(inputchannels, outputchannels; verbose = false)

Starts a task to monitor all the input_channels to receive a signal and performs synchronisation of all the terms in the input channels.

The gradients from these channels are expected to be of the form of a NamedTuple as produced by Zygote. A typical example would be

julia> resnet = ResNet(); # from Metalhead.jl or could be any model we wish to train

julia> using Zygote

julia> gs, _ = gradient(resnet, rand(Float32, 224, 224, 3, 1)) do m, x
         sum(m(x))
       end;

gs is what would be sent in the channels from every worker.

All the input channels are expected to be started at the remote processes with a size 1 so only one gradient may be published at one time. The output channels are similar with the channel being started on the process where the sync in expected to happen.

Typical configuration would look like:

input_channels = (RemoteChannel(() -> Channel(1), p) for p in workers())
output_channels = (RemoteChannel(() -> Channel(1), 1) for p in workers())

Set verbose = true to get more detailed information as the synchronisation happens.

Note that syncgrads currently requires serialization of gradients from every device with every iteration of the data loader. This is inefficient and has been surpassed with techniques involving "data layers" such as Nvidia NCCL or UCX which work to perform reduction over several GPUs in a better optimised manner. This is under developement in the Julia ecosystem as well.

High Level training pipeline

A basic pipeline would look very similar to how Flux.train! functions. We need a cost function, some data and parameters and we are off to the races. There are some subtlties however. Rather than using the gradients that every copy of the model produces with the data it used, data parallel training does not directly use these gradients. Instead, the gradients from every model have to be reduced together to average it out to get the effect of training over all the data used to perform a single training step. This amounts to:

# state => state of optimiser
# model => model to be trained
# opt => optimiser
# data => collection of data
for d in data
  x, y = d  # x => input data, y => labels
  gs = gradient(model) do m
    loss(m(x), y)
  end

  # note the pseudo code call in the next line
  @synchronize_gradients_over_all_GPUs
  model, state = opt(model, updated_grads, state)
end

This looks very similar to the typical supervised learning training loop from Flux.

In fact, it is! With the addition of the synchrnisation part, we can also extend it to several forms of semi-supervised and unsupervised learning scenarios. This is part of the future work of this pacakge, and something actively being researched in the Julia community and elsewhere.

ResNetImageNet.start — Function

start(loss, data_tree, key, model,
      input_channels, output_channels;
      class_idx,
      verbose = false,
      opt_args = (),
      opt = Optimisers.ADAM(),
      kwargs...)

The high level function that performs training of the model over the specified data and configuration.

loss: Typical loss function used to optimise the model with the data. It is fed all the data that every iteration of a DataLaader produces, such that the first element of the produced data is first send to the model. The calling signature of the loss looks like:

x, y, z... = iterate(dataloader) loss(model(x), y, z...)

data_tree: the data tree that one would associate with a dataset described by DataSets.dataset.

key: the key to the training data. Typically the LOC_train__solutions.csv` for the case of ILSVRC.

model: model to be trained

inputchannels, outputchannels: See syncgrad for details

Keyword Arguments:

This is a non-exhaustive list of keyword arguments currently supported.

class_idx: a list of the labels to be trained on, a Vector, or Base.OneTo

o: The type of optimiser to use. Optimisers.jl provides a number of supported optimisers.

opt_args: A tuple containing arguments to the optimiser as described by o

opt: The output of o(opt_args...). Useful to provide initial optimisers. Can also be associated with schedulers and decays as required.

cycles: the number of times the dataset is sampled in a single call to start.

batchsize: the number of observations per batch per GPU.

nsamples: The number of datapoints to be sampled at once. This subset of data is loaded by every process independently and creates a DataLoader from it.

sts: A vector of length nworkers() which each contain the current state of the optimiser. This is initialised as [Optimisers.state(opt, model) for _ in workers()]

saveweights: Defaults to false, set to true to save training checkpoints

vals: A vector of validation sets to be used as validation sets while training. These also add logging statements while training. Disable with (and defaults to) [nothing for _ in workers()].

workers: A list of processes used to train the model

devices: A DeviceSet() or iterable of Device(); used to target the GPU used to train.

verbose: Defaults to false. Set to true to enable logging of helpful information while debugging and training.